Refactor scripts to reproduce data.
