+++ COMET/README.md
@@ -74,7 +74,8 @@ seg_scores, sys_score = model.predict(data, batch_size=8, gpus=1)
 ```
 
 ## Reproduce
-Follow the [instructions](reproduce/) to reproduce the Doc-COMET(-QE) results from the paper.
+To reproduce the Doc-COMET results from the paper run the [score_doc-metrics.py](/score_doc-metrics.py) script with the flags `--model comet` and `--doc`. For the Doc-COMET-QE results also add the `--qe` flag.
+
 
 ## Paper
